+++ lib/activeadmin_mcp/resource_registry.rb
@@ -4,14 +4,28 @@ module ActiveadminMcp
   module ResourceRegistry
     class << self
       def all
-        discover.map { |r| resource_info(r) }
+        resources.map { |entry| resource_info(entry) }
+      end
+
+      def resources
+        discover.map { |r| entry(r) }
       end
 
       def find(name)
-        resource = discover.find { |r| r.resource_class.name == name }
-        return unless resource
+        resources.find { |entry| entry[:name] == name }
+      end
 
-        { name: resource.resource_class.name, model: resource.resource_class, config: resource }
+      def resource_info(entry)
+        klass = entry[:model]
+        {
+          name: klass.name,
+          table: klass.table_name,
+          attributes: klass.column_names - sensitive_attributes,
+        }
+      end
+
+      def sensitive_attributes
+        %w[encrypted_password password_digest reset_password_token api_key secret]
       end
 
       private
@@ -26,17 +40,8 @@ module ActiveadminMcp
         end || []
       end
 
-      def resource_info(resource)
-        klass = resource.resource_class
-        {
-          name: klass.name,
-          table: klass.table_name,
-          attributes: klass.column_names - sensitive_attributes,
-        }
-      end
-
-      def sensitive_attributes
-        %w[encrypted_password password_digest reset_password_token api_key secret]
+      def entry(resource)
+        { name: resource.resource_class.name, model: resource.resource_class, config: resource }
       end
     end
   end